data_retrieval/requirements.txt
flask
kafka-python
confluent_kafka
mysql.connector
